netplay: Move ban list code to netpermissions.h/.cpp
